feat(strict-store): forbid passing non-enum keys, even if they are registered in registry.
This is a deliberate choice. If you have enum with missing key and use it in StrictStore, it will display errors `.on` and `.notify`, even if it is technically registered. as far as I can, this is the best I can do to prevent foreign keys from being **registered and used**.